chore: update linux setup script to python3

+++ linux/packages/docker.sh
@@ -1,4 +1,13 @@
 #!/usr/bin/env bash
 set -euo pipefail
-curl -fsSL https://get.docker.com -o get-docker.sh
-sudo sh get-docker.sh && rm -rf get-docker.sh
+
+echo "📦 Installing docker CE Server..."
+
+if command -v docker &>/dev/null; then
+  echo "docker is already installed"
+else
+  curl -fsSL https://get.docker.com -o get-docker.sh
+  sudo sh get-docker.sh && rm -rf get-docker.sh
+  echo "Create default network.."
+  docker network create web_network
+fi
